enable POST-to-redirect at /auth/token
At this proxy endpoint, if the user supplies a token in the
access_token param, it will be used for auth and then (assuming
success) they will be redirected to the "redirect" parameter.

This way an external UI could pass the token to the proxy via the
browser in a way that doesn't go in the browser history or proxy
logs, and have it persisted in the browser session.